pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nats-connect-opts"
authors = [
    { name = "Guillaume Charbonnier", email = "gu.charbon@gmail.com" },
]
description = "Opinionated way to configure NATS connect options in Python"
readme = "README.md"
requires-python = ">=3.8,<4.0"
classifiers = [
    'Intended Audience :: Developers',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    "License :: OSI Approved :: Apache Software License",
]
dynamic = ["version"]
dependencies = ["nats-py"]

[project.optional-dependencies]
build = ["pip-tools", "build", "wheel"]
dev = [
    "black",
    "isort",
    "invoke",
    "flake8",
    "mutmut",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "nats-test-server",
]
docs = [
    "requests",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-material",
    "mkdocs-section-index",
    "mkdocstrings[python]",
    "pymdown-extensions",
    "termynal",
]

[tool.setuptools.dynamic]
version = { attr = "nats_contrib.connect_opts.__about__.__version__" }

[tool.pyright]
pythonVersion = "3.8"
include = ["tests", "src"]
strict = ["src"]
exclude = ["**/.venv", "**/node_modules", "**/__pycache__", ".git", "**/build"]
venv = ".venv"
venvPath = "."
typeCheckingMode = "basic"
reportUnnecessaryTypeIgnoreComment = "warning"
